+++ lib/salah.rb
@@ -1,14 +1,13 @@
 class Salah
   require 'openstruct-from_hash'
   require_relative 'salah/prayer'
+  require_relative 'salah/date'
   require_relative 'salah/school'
   require_relative 'salah/location'
   require_relative 'salah/juristic'
   require_relative 'salah/http'
   require_relative 'salah/response'
   require_relative 'salah/version'
-
-  Date = Struct.new(:iso8601, :hijri, :timestamp)
 
   #
   # @return [Array<Salah::Juristic>]